modified the delete function on controller

+++ app/Http/Controllers/DailyTransactionsController.php
@@ -7,11 +7,10 @@ use Illuminate\Validation\ValidationException;
 use Illuminate\Database\QueryException;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use App\Models\Daily_transactions as Transactions;
-use App\Models\items;
 use Illuminate\Support\Facades\DB;
 use Carbon\Carbon;
 
-use function PHPUnit\Framework\isEmpty;
+
 
 class DailyTransactionsController extends Controller
 {
@@ -288,6 +287,15 @@ class DailyTransactionsController extends Controller
             );
 
             $transactions = Transactions::where('id', $validatedID['id'])->firstOrFail();
+
+             // Check if transaction is from today
+        if (!Carbon::parse($transactions->transaction_date)->isToday()) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Cannot delete transactions from previous days. Only today\'s transactions can be deleted.',
+            ], 403); // 403 Forbidden status code
+        }
+
             $transactions->delete();
             return response()->json([
                 'success' => true,
